Doc fix + lowercase pkg name on Debian
Fix the documentation of perl::module to state that use_package is
"false" by default.

On Debian based systems, the $pkg_name is not correctly determined by
perl::module. If a manifest defines a perl::module for "Net::DNS",
and $use_package is "true", then it defines a package resource for
"libNet-DNS-perl".  The underlying APT system is smart enough to
recognise this, but Puppet is not clever enough to see that
libnet-dns-perl and libNet-DNS-perl are basically identical. As a
result, every Puppet agent run will result in a notice for
Package[libNet-DNS-perl] stating "ensure changed 'purged' to
'present'".