Fix OS Family Matching for Future Parser
The future parser for Puppet 4 doesn't like these operating system names to be unquoted. This causes the match to fall through to the `default` case, which doesn't define `$package_downcase`, which causes the following error:

```
Error: Evaluation Error: No matching entry for selector parameter with value '' at /vagrant/modules/perl/manifests/module.pp:34:15 on node
 stg-dbmaster1
```

If I resolve the `$osfamily` fact, it resolves to `RedHat`, so it _should_ match, but it doesn't and falls through to the `default` case. As soon as I wrap the above operating system family names in params.pp with quotes, execution completes as expected.